ai_agent/requirements.txt
langgraph>=0.2.0
langchain-google-genai>=2.0.0
polars>=0.20.0
duckdb>=0.10.0
pandas>=2.0.0
python-dotenv>=1.0.0
gradio>=4.0.0
